change uint64 to uint32

+++ apps/arena/lib/arena/serialization/messages.pb.ex
@@ -130,7 +130,7 @@ defmodule Arena.Serialization.GameFinished.PlayersEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -153,7 +153,7 @@ defmodule Arena.Serialization.PingUpdate do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:latency, 1, type: :uint64)
+  field(:latency, 1, type: :uint32)
 end
 
 defmodule Arena.Serialization.GameJoined do
@@ -161,7 +161,7 @@ defmodule Arena.Serialization.GameJoined do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:player_id, 1, type: :uint64, json_name: "playerId")
+  field(:player_id, 1, type: :uint32, json_name: "playerId")
   field(:config, 2, type: Arena.Serialization.Configuration)
   field(:bounties, 3, repeated: true, type: Arena.Serialization.BountyInfo)
 end
@@ -213,8 +213,8 @@ defmodule Arena.Serialization.ConfigCharacter do
   field(:active, 2, type: :bool)
   field(:base_speed, 3, type: :float, json_name: "baseSpeed")
   field(:base_size, 4, type: :float, json_name: "baseSize")
-  field(:base_health, 5, type: :uint64, json_name: "baseHealth")
-  field(:max_inventory_size, 6, type: :uint64, json_name: "maxInventorySize")
+  field(:base_health, 5, type: :uint32, json_name: "baseHealth")
+  field(:max_inventory_size, 6, type: :uint32, json_name: "maxInventorySize")
 
   field(:skills, 7,
     repeated: true,
@@ -237,19 +237,19 @@ defmodule Arena.Serialization.ConfigServerUpdate do
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
   field(:timestamp_difference_samples_to_check_warning, 1,
-    type: :uint64,
+    type: :uint32,
     json_name: "timestampDifferenceSamplesToCheckWarning"
   )
 
   field(:timestamp_differences_samples_max_length, 2,
-    type: :uint64,
+    type: :uint32,
     json_name: "timestampDifferencesSamplesMaxLength"
   )
 
-  field(:show_warning_threshold, 3, type: :uint64, json_name: "showWarningThreshold")
-  field(:stop_warning_threshold, 4, type: :uint64, json_name: "stopWarningThreshold")
-  field(:ms_without_update_show_warning, 5, type: :uint64, json_name: "msWithoutUpdateShowWarning")
-  field(:ms_without_update_disconnect, 6, type: :uint64, json_name: "msWithoutUpdateDisconnect")
+  field(:show_warning_threshold, 3, type: :uint32, json_name: "showWarningThreshold")
+  field(:stop_warning_threshold, 4, type: :uint32, json_name: "stopWarningThreshold")
+  field(:ms_without_update_show_warning, 5, type: :uint32, json_name: "msWithoutUpdateShowWarning")
+  field(:ms_without_update_disconnect, 6, type: :uint32, json_name: "msWithoutUpdateDisconnect")
 end
 
 defmodule Arena.Serialization.ConfigSkill do
@@ -258,12 +258,12 @@ defmodule Arena.Serialization.ConfigSkill do
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
   field(:name, 1, type: :string)
-  field(:cooldown_ms, 2, type: :uint64, json_name: "cooldownMs")
-  field(:execution_duration_ms, 3, type: :uint64, json_name: "executionDurationMs")
+  field(:cooldown_ms, 2, type: :uint32, json_name: "cooldownMs")
+  field(:execution_duration_ms, 3, type: :uint32, json_name: "executionDurationMs")
   field(:targetting_radius, 4, type: :float, json_name: "targettingRadius")
   field(:targetting_angle, 5, type: :float, json_name: "targettingAngle")
   field(:targetting_range, 6, type: :float, json_name: "targettingRange")
-  field(:stamina_cost, 7, type: :uint64, json_name: "staminaCost")
+  field(:stamina_cost, 7, type: :uint32, json_name: "staminaCost")
 end
 
 defmodule Arena.Serialization.GameState.PlayersEntry do
@@ -271,7 +271,7 @@ defmodule Arena.Serialization.GameState.PlayersEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -280,7 +280,7 @@ defmodule Arena.Serialization.GameState.ProjectilesEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -289,7 +289,7 @@ defmodule Arena.Serialization.GameState.PlayerTimestampsEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: :int64)
 end
 
@@ -298,8 +298,8 @@ defmodule Arena.Serialization.GameState.DamageTakenEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
-  field(:value, 2, type: :uint64)
+  field(:key, 1, type: :uint32)
+  field(:value, 2, type: :uint32)
 end
 
 defmodule Arena.Serialization.GameState.DamageDoneEntry do
@@ -307,8 +307,8 @@ defmodule Arena.Serialization.GameState.DamageDoneEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
-  field(:value, 2, type: :uint64)
+  field(:key, 1, type: :uint32)
+  field(:value, 2, type: :uint32)
 end
 
 defmodule Arena.Serialization.GameState.PowerUpsEntry do
@@ -316,7 +316,7 @@ defmodule Arena.Serialization.GameState.PowerUpsEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -325,7 +325,7 @@ defmodule Arena.Serialization.GameState.ItemsEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -334,7 +334,7 @@ defmodule Arena.Serialization.GameState.ObstaclesEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -343,7 +343,7 @@ defmodule Arena.Serialization.GameState.PoolsEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -352,7 +352,7 @@ defmodule Arena.Serialization.GameState.CratesEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -361,7 +361,7 @@ defmodule Arena.Serialization.GameState.BushesEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -370,7 +370,7 @@ defmodule Arena.Serialization.GameState.TrapsEntry do
 
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:key, 1, type: :uint64)
+  field(:key, 1, type: :uint32)
   field(:value, 2, type: Arena.Serialization.Entity)
 end
 
@@ -443,14 +443,14 @@ defmodule Arena.Serialization.Entity do
 
   oneof(:aditional_info, 0)
 
-  field(:id, 1, type: :uint64)
+  field(:id, 1, type: :uint32)
   field(:category, 2, type: :string)
   field(:shape, 3, type: :string)
   field(:name, 4, type: :string)
   field(:position, 5, type: Arena.Serialization.Position)
   field(:radius, 6, type: :float)
   field(:vertices, 7, repeated: true, type: Arena.Serialization.Position)
-  field(:collides_with, 8, repeated: true, type: :uint64, json_name: "collidesWith")
+  field(:collides_with, 8, repeated: true, type: :uint32, json_name: "collidesWith")
   field(:speed, 9, type: :float)
   field(:direction, 10, type: Arena.Serialization.Direction)
   field(:is_moving, 11, type: :bool, json_name: "isMoving")
@@ -471,7 +471,7 @@ defmodule Arena.Serialization.Player.CooldownsEntry do
   use Protobuf, map: true, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
   field(:key, 1, type: :string)
-  field(:value, 2, type: :uint64)
+  field(:value, 2, type: :uint32)
 end
 
 defmodule Arena.Serialization.Player do
@@ -479,8 +479,8 @@ defmodule Arena.Serialization.Player do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:health, 1, type: :uint64)
-  field(:kill_count, 2, type: :uint64, json_name: "killCount")
+  field(:health, 1, type: :uint32)
+  field(:kill_count, 2, type: :uint32, json_name: "killCount")
 
   field(:current_actions, 3,
     repeated: true,
@@ -488,16 +488,16 @@ defmodule Arena.Serialization.Player do
     json_name: "currentActions"
   )
 
-  field(:available_stamina, 4, type: :uint64, json_name: "availableStamina")
-  field(:max_stamina, 5, type: :uint64, json_name: "maxStamina")
-  field(:stamina_interval, 6, type: :uint64, json_name: "staminaInterval")
+  field(:available_stamina, 4, type: :uint32, json_name: "availableStamina")
+  field(:max_stamina, 5, type: :uint32, json_name: "maxStamina")
+  field(:stamina_interval, 6, type: :uint32, json_name: "staminaInterval")
   field(:recharging_stamina, 7, type: :bool, json_name: "rechargingStamina")
   field(:character_name, 8, type: :string, json_name: "characterName")
-  field(:power_ups, 9, type: :uint64, json_name: "powerUps")
+  field(:power_ups, 9, type: :uint32, json_name: "powerUps")
   field(:effects, 10, repeated: true, type: Arena.Serialization.Effect)
   field(:inventory, 11, type: Arena.Serialization.Item)
   field(:cooldowns, 12, repeated: true, type: Arena.Serialization.Player.CooldownsEntry, map: true)
-  field(:visible_players, 13, repeated: true, type: :uint64, json_name: "visiblePlayers")
+  field(:visible_players, 13, repeated: true, type: :uint32, json_name: "visiblePlayers")
   field(:on_bush, 14, type: :bool, json_name: "onBush")
   field(:forced_movement, 15, type: :bool, json_name: "forcedMovement")
 end
@@ -524,8 +524,8 @@ defmodule Arena.Serialization.Projectile do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:damage, 1, type: :uint64)
-  field(:owner_id, 2, type: :uint64, json_name: "ownerId")
+  field(:damage, 1, type: :uint32)
+  field(:owner_id, 2, type: :uint32, json_name: "ownerId")
   field(:status, 3, type: Arena.Serialization.ProjectileStatus, enum: true)
   field(:skill_key, 4, type: :string, json_name: "skillKey")
 end
@@ -543,7 +543,7 @@ defmodule Arena.Serialization.PowerUp do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:owner_id, 1, type: :uint64, json_name: "ownerId")
+  field(:owner_id, 1, type: :uint32, json_name: "ownerId")
   field(:status, 2, type: Arena.Serialization.PowerUpstatus, enum: true)
 end
 
@@ -552,8 +552,8 @@ defmodule Arena.Serialization.Crate do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:health, 1, type: :uint64)
-  field(:amount_of_power_ups, 2, type: :uint64, json_name: "amountOfPowerUps")
+  field(:health, 1, type: :uint32)
+  field(:amount_of_power_ups, 2, type: :uint32, json_name: "amountOfPowerUps")
   field(:status, 3, type: Arena.Serialization.CrateStatus, enum: true)
 end
 
@@ -562,7 +562,7 @@ defmodule Arena.Serialization.Pool do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:owner_id, 1, type: :uint64, json_name: "ownerId")
+  field(:owner_id, 1, type: :uint32, json_name: "ownerId")
 end
 
 defmodule Arena.Serialization.Bush do
@@ -576,7 +576,7 @@ defmodule Arena.Serialization.Trap do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:owner_id, 1, type: :uint64, json_name: "ownerId")
+  field(:owner_id, 1, type: :uint32, json_name: "ownerId")
   field(:name, 2, type: :string)
   field(:status, 3, type: Arena.Serialization.TrapStatus, enum: true)
 end
@@ -587,7 +587,7 @@ defmodule Arena.Serialization.PlayerAction do
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
   field(:action, 1, type: Arena.Serialization.PlayerActionType, enum: true)
-  field(:duration, 2, type: :uint64)
+  field(:duration, 2, type: :uint32)
   field(:destination, 3, type: Arena.Serialization.Position)
   field(:direction, 4, type: Arena.Serialization.Position)
 end
@@ -622,7 +622,7 @@ defmodule Arena.Serialization.UseItem do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:item, 1, type: :uint64)
+  field(:item, 1, type: :uint32)
 end
 
 defmodule Arena.Serialization.SelectBounty do
@@ -698,8 +698,8 @@ defmodule Arena.Serialization.KillEntry do
 
   use Protobuf, syntax: :proto3, protoc_gen_elixir_version: "0.12.0"
 
-  field(:killer_id, 1, type: :uint64, json_name: "killerId")
-  field(:victim_id, 2, type: :uint64, json_name: "victimId")
+  field(:killer_id, 1, type: :uint32, json_name: "killerId")
+  field(:victim_id, 2, type: :uint32, json_name: "victimId")
 end
 
 defmodule Arena.Serialization.BountyInfo do
